rename highlighter
