+++ packages/vue-instantsearch/src/util/vue-compat/Highlighter/Highlighter-vue2.vue
