+++ packages/vue-instantsearch/src/util/vue-compat/Highlighter/index-vue2.js
@@ -1,1 +1,1 @@
-export { default } from './Highlighter.vue';
+export { default } from './Highlighter-vue2.vue';
